requirements.txt
# Core dependencies
scanpy>=1.9.3
anndata>=0.9.0
scvelo>=0.2.5
dynamo-release>=1.3.0

# Deep learning
torch>=2.0.0
torchdiffeq>=0.2.3

# Scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-misc>=0.1.4

# Dimensionality reduction
umap-learn>=0.5.3

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Gene set enrichment
gseapy>=1.0.4

# Web interface
streamlit>=1.28.0

# Utilities
tqdm>=4.65.0
h5py>=3.8.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0